+++ R/splineCoxreg2.R
@@ -31,11 +31,11 @@
 #' @return A list containing the following components:
 #'   \item{model}{A character string indicating the shape of the baseline hazard function used.}
 #'   \item{parameter}{A numeric vector of the parameters defining the baseline hazard shape.}
-#'   \item{beta}{A named vector with the estimates, standard errors, and 95% confidence intervals for the regression coefficients}
-#'   \item{gamma}{A named vector with the estimate, standard error, and 95% confidence interval for the baseline hazard parameter}
+#'   \item{beta}{A named vector with the estimates, standard errors, and 95\% confidence intervals for the regression coefficients.}
+#'   \item{gamma}{A named vector with the estimate, standard error, and 95\% confidence interval for the baseline hazard parameter.}
 #'   \item{loglik}{A named vector containing the log-likelihood (\code{LogLikelihood}),
 #'                 Akaike Information Criterion (\code{AIC}), and Bayesian Information
-#'                 Criterion (\code{BIC}) for the best-fitting model}
+#'                 Criterion (\code{BIC}) for the best-fitting model.}
 #'   \item{other_models}{A data frame containing the log-likelihood (\code{LogLikelihood}) for all other evaluated models,
 #'                            with model names as row names.}
 #'   \item{plot}{A baseline hazard function plot for the best-fitting model (if \code{plot = TRUE}).}
@@ -50,7 +50,11 @@
 #' M = c("constant", "increase", "decrease")
 #' reg2 <- splineCox.reg2(t.event, event, Z, model = M)
 #' print(reg2)
-#'
+#' @references
+#' Teranishi, R.; Furukawa, K.; Emura, T. (2025).
+#' A Two-Stage Estimation Approach to Cox Regression Under the Five-Parameter Spline Model
+#' Mathematics 13(4), 616. \doi{10.3390/math13040616}
+#' Available at \url{https://www.mdpi.com/2227-7390/13/4/616}
 splineCox.reg2 <- function (t.event, event, Z, xi1 = min(t.event), xi3 = max(t.event),
                             model = names(shape.list), p0 = rep(0, 1 + ncol(as.matrix(Z))),
                             plot = TRUE)